move InMemoryOrderBook into paper_client
